+++ CHANGELOG.md
@@ -18,12 +18,16 @@ Changelog since 1.3.3:
 ### Changed
  - [Kraken X42/X52/X62/X72] Increase resolution of fan and pump profiles
  - [extra/krakencurve-poc] Refresh syntax and sensor names; get CPU temperature on macOS with iStats
- - Dynamically find the HID interface and endpoints of PyUsbHids
+ - Use hidapi for HIDs on Windows
  - Improve the documentation
 ### Fixed
  - [HUE 2] Add missing identifiers for some HUE2 accessories (#95; #109)
  - [NZXT E500/E650/E850] Fix CAM-like decoding of firmware version (#46, comment)
  - [HUE 2] Select the lighting channel with a bitmask (#109)
+### Deprecated
+ - Deprecate and ignore `--hid` override HID API selection
+### Removed
+ - Remove the PyUsbHid device backend for HIDs
 
 
 ## [1.3.3] – 2020-02-18